The applying of TDB Poison (Stun) moved in order to prevent the Poison triggering even if you miss.
There are two additional parameters that can be used (I'm not sure what other spells have a double hit that lead to an effect on the second hit)

Just be sure to set the parameters on the DelayedAction initialisation.

I've added a summary to the DelayedType.Damage to make it easier.

Suggest adding a new class or a struct to hold delayed attack information rather than using parameters for better control.