[BUG] - Colour not correctly updating
When you war a guild, yourself and the enemy turn blue. This is only corrected when you re-log but everyone is visible as blue whom is at war, even to people who are not currently war'd.

This prevents name colour refreshes from correctly applying, as it defaults to this colour causing major confusion.