Observe button now displayed
Observe Button is now being displayed on the CharacterInspectWindow. I've moved the bottom buttons left slightly to all the Observe button to fit neatly.

I've set the button to be displayed regardless of if the Character has observe enabled/disabled but this wont bypass the AllowObserve check so is only a visual change.